requirements.txt
amqp==2.3.2
beautifulsoup4==4.5.0
billiard==3.5.0.4
certifi==2018.4.16
chardet==3.0.4
html5lib==1.0.1
idna==2.7
kombu==4.2.1
lxml==4.4.2
numpy==1.17.4
pandas==0.25.3
pip==19.3.1
pymssql==2.1.4
pyodbc==4.0.27
pyperclip==1.7.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
schedule==0.6.0
setuptools==40.8.0
six==1.11.0
SQLAlchemy==1.3.11
urllib3==1.25.7
vine==1.1.4
webencodings==0.5.1
